requirements.txt
python-telegram-bot>=20.0
guessit
hachoir
requests
motor
Flask